Make etcd.conf editing more robust
The AWK incantation that we used to remove an existing /usr/bin/etcd
invocation from /etc/init/etcd.conf did not work correctly if there
was nothing in the file after that invocation; it left the final line
of the invocation in place.  This commit simplifies and improves the
script so that it works in that case.

Change-Id: Icc82ed36485b24468b02eabddac86ae73ea6a29b